requirements.txt
requests>=2.31.0
qrcode>=7.4.2
Pillow>=10.0.0
schedule>=1.2.1
python-dotenv>=1.0.0
pycryptodome>=3.19.0